added displayOptions type to MiddlewareOptions
