+++ src/utils/types.ts
@@ -2,3 +2,12 @@ export type WorkerCallback = (
   path: string,
   relative: boolean,
 ) => Promise<Worker>;
+
+export interface VoyagerDisplayOptions {
+  rootType?: string;
+  skipRelay?: boolean;
+  skipDeprecated?: boolean;
+  showLeafFields?: boolean;
+  sortByAlphabet?: boolean;
+  hideRoot?: boolean;
+}
